% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFinalCols.R
\name{assignFinalCols}
\alias{assignFinalCols}
\title{Assign values to final preprocessed column names}
\usage{
assignFinalCols(data, noise_reduction, ...)
}
\arguments{
\item{data}{dataframe}

\item{noise_reduction}{Boolean indicating if noise reduction should be implemented on gaze data}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
data
}
\description{
Assign values to final preprocessed column names
}
