% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_preprocess.R
\name{batch_preprocess}
\alias{batch_preprocess}
\title{batch_preprocess is used to run the preprocessing function for a set of files in a BIDS like directory. This will process files in parallel, and compute total runtime to process.}
\usage{
batch_preprocess(directoryBIDS, batch_name, num_cores = NULL, ...)
}
\arguments{
\item{directoryBIDS}{filepath to the BIDS directory of ET data to process}

\item{batch_name}{string label for what to call the batch. This will be added to the output files}

\item{num_cores}{optional parameter to specify number of cores to use. If not specified, function will use 80\\% of available cores}

\item{...}{additional parameters}
}
\value{
data
}
\description{
batch_preprocess is used to run the preprocessing function for a set of files in a BIDS like directory. This will process files in parallel, and compute total runtime to process.
}
