% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHz.R
\name{calcHz}
\alias{calcHz}
\title{calcHz
calculate eye tracker temporal resolution (Hz)}
\usage{
calcHz(data, timepointIDs = 1:nrow(data))
}
\arguments{
\item{data}{a dataframe, with recording timestamps IN MILLISECONDS}

\item{timepointIDs}{a range of timepoints by timepoint ID to include. Default includes all timepoints}
}
\value{
recHz
}
\description{
calcHz
calculate eye tracker temporal resolution (Hz)
}
\examples{
exampledata <- data.frame("recordingTimestamp_ms" = seq(1,1000,2))
calcHz(exampledata)
calcHz(exampledata, 2:nrow(exampledata)) #excludes the first timepoint

}
