% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFrequency_hz.R
\name{calculateFrequency_hz}
\alias{calculateFrequency_hz}
\title{calculateFrequency_hz
calculate eye tracker temporal resolution (Hz)}
\usage{
calculateFrequency_hz(data, timepointIDs = 1:nrow(data))
}
\arguments{
\item{data}{a dataframe, with recording timestamps IN MILLISECONDS}

\item{timepointIDs}{a range of timepoints by timepoint ID to include. Default includes all timepoints}
}
\value{
recordingFrequency_hz
}
\description{
calculateFrequency_hz
calculate eye tracker temporal resolution (Hz)
}
\examples{
exampledata <- data.frame("recordingTimestamp_ms" = seq(1, 1000, 2))
calculateFrequency_hz(exampledata)
calculateFrequency_hz(exampledata, 2:nrow(exampledata)) # excludes the first timepoint

}
