% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGaze_va.R
\name{calculateGaze_va}
\alias{calculateGaze_va}
\title{calculate visual angle from pixels}
\usage{
calculateGaze_va(
  data,
  displayResolutionX_px,
  displayResolutionY_px,
  displayDimensionX_mm,
  displayDimensionY_mm,
  gazeX,
  gazeY,
  distanceZ,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{displayResolutionX_px}{int display resolution for width in px}

\item{displayResolutionY_px}{int display resolution for height in px}

\item{displayDimensionX_mm}{int display width in mm}

\item{displayDimensionY_mm}{int display height in mm}

\item{gazeX}{string column name of gazepoints on X axis}

\item{gazeY}{string column name of gazepoints on Y axis}

\item{distanceZ}{string column name of distance from screen for each timepoint}

\item{...}{additional passed parameters from parent function}
}
\value{
data_va dataframe with calculated visual angles in columns appended _va
}
\description{
calculate visual angle from pixels
}
