% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_difference.R
\name{calculateTimeDifference}
\alias{calculateTimeDifference}
\title{calculateTimeDifference - get current time}
\usage{
calculateTimeDifference(start, stop, units = "secs")
}
\arguments{
\item{start}{timestamp for start time}

\item{stop}{timestamp for stop time}

\item{units}{units difference should be returned, (secs, mins, hours, etc)}
}
\value{
timestamp in seconds, from sys.time()
}
\description{
calculateTimeDifference - get current time
}
