% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVelocity_va_ms.R
\name{calculateVelocity_va_ms}
\alias{calculateVelocity_va_ms}
\title{Calculate Velocity in Visual Angle}
\usage{
calculateVelocity_va_ms(data, gazeX_va, gazeY_va, timestamp, ...)
}
\arguments{
\item{data}{dataframe}

\item{gazeX_va}{string, column containing x coordinates of gaze stream in visual angles}

\item{gazeY_va}{string, column containing y coordinates of gaze stream in visual angles}

\item{timestamp}{string, column containing recording timestamps in ms}

\item{...}{additional passed parameters from parent function}
}
\value{
data a dataframe with calculated velocity for gaze X and gaze Y and euclidean distance
}
\description{
Establish timestamps and
}
