% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVA.r
\name{calcVA}
\alias{calcVA}
\title{Title}
\usage{
calcVA(
  data,
  display.resx,
  display.resy,
  display.dimx,
  display.dimy,
  gazeX,
  gazeY,
  distZ,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{display.resx}{int display resolution for width in px}

\item{display.resy}{int display resolution for height in px}

\item{display.dimx}{int display width in mm}

\item{display.dimy}{int display height in mm}

\item{gazeX}{string column name of gazepoints on X axis}

\item{gazeY}{string column name of gazepoints on Y axis}

\item{distZ}{string column name of distance from screen for each timepoint}

\item{...}{additional passed parameters from parent function}
}
\value{
data dataframe with calculated visual angles in columns appended .va
}
\description{
Title
}
