% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVel.r
\name{calcVel}
\alias{calcVel}
\title{Calculate Velocity}
\usage{
calcVel(data, xcoord, ycoord, timestamp, ...)
}
\arguments{
\item{data}{dataframe}

\item{xcoord}{string, column containing x coordinates of gaze stream, converted to visual angle}

\item{ycoord}{string, column containing y coordinates of gaze stream, converted to visual angle}

\item{timestamp}{string, column containing recording timestamps in ms}

\item{...}{additional passed parameters from parent function}
}
\value{
data a dataframe with calculated velocity for gaze X and gaze Y and euclidean distance
}
\description{
Establish timestamps and
}
