% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGazeData.R
\name{checkGazeData}
\alias{checkGazeData}
\title{Check that file contains any gaze data to be processed}
\usage{
checkGazeData(data, gazeCol, ...)
}
\arguments{
\item{data}{a dataframe}

\item{gazeCol}{column containing gaze data to assess}

\item{...}{additional passed parameters from parent function}
}
\value{
dataExists a Boolean indicating if any valid gaze data exists
}
\description{
Check that file contains any gaze data to be processed
}
