% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTimestamps.R
\name{checkTimestamps}
\alias{checkTimestamps}
\title{Check that data timestamps are in chronological order}
\usage{
checkTimestamps(data, timestamps)
}
\arguments{
\item{data}{a dataframe}

\item{timestamps}{string indicating variable containing time course data to be checked}
}
\value{
logical indicating TRUE if timestamps are chronologically ordered, FALSE if disordered
}
\description{
Check that data timestamps are in chronological order
}
