% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyGazeIVT.R
\name{classifyGazeIVT}
\alias{classifyGazeIVT}
\title{Apply IVT Classifier}
\usage{
classifyGazeIVT(
  data,
  velocity,
  gazeX_va,
  gazeY_va,
  recordingFrequency_hz,
  fixationVelocityThreshold = 50,
  maxAdjacentFixationAngle = 0.5,
  maxAdjacentFixationTime = 75,
  minFixationDuration = 60,
  ...
)
}
\arguments{
\item{data}{a dataframe}

\item{velocity}{string - variable column name containing velocity (in visual angles)}

\item{gazeX_va}{string - variable column name containing X gaze coordinates in visual angles}

\item{gazeY_va}{string - variable column name containing Y gaze coordinates in visual angles}

\item{recordingFrequency_hz}{recording frequency in Hz}

\item{fixationVelocityThreshold}{fixation classifier velocity threshold in VA/s. Default: 50 VA / s}

\item{maxAdjacentFixationAngle}{maximum distance (in VA) between two fixations to be merged. Default: 0/5 VA}

\item{maxAdjacentFixationTime}{maximum time lapsed (in ms) between two fixations to be merged. Default: 75 ms}

\item{minFixationDuration}{threshold for minimum duration of a given fixation to be considered acceptable. Default: 60 ms}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
outputs_dfs
}
\description{
Apply IVT Classifier
}
