% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyOffscreen.R
\name{classifyOffscreen}
\alias{classifyOffscreen}
\title{Classify Offscreen gaze points in preprocessing pipeline}
\usage{
classifyOffscreen(data, display.resx, display.resy, eyeSelection, ...)
}
\arguments{
\item{data}{dataframe}

\item{display.resx}{X display resolution in pixels}

\item{display.resy}{Y display resolution in pixels}

\item{eyeSelection}{"Maximize", "Strict", "Left", "Right"}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
data
}
\description{
Classify Offscreen gaze points in preprocessing pipeline
}
