% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyOffscreenGaze.R
\name{classifyOffscreenGaze}
\alias{classifyOffscreenGaze}
\title{Classify Offscreen gaze points in preprocessing pipeline}
\usage{
classifyOffscreenGaze(
  data,
  displayResolutionX_px,
  displayResolutionY_px,
  eyeSelection_method,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{displayResolutionX_px}{X display resolution in pixels}

\item{displayResolutionY_px}{Y display resolution in pixels}

\item{eyeSelection_method}{"Maximize", "Strict", "Left", "Right"}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
data
}
\description{
Classify Offscreen gaze points in preprocessing pipeline
}
