% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.r
\name{denoise}
\alias{denoise}
\title{De-noise and smooth rolling averages}
\usage{
denoise(
  data,
  recHz,
  input_cols,
  smooth_window_ms = 20,
  noise_reduction = TRUE,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{recHz}{recording Hz}

\item{input_cols}{list of columns to smooth. Recommended to smooth gaze, pupil, and Z Dist columns}

\item{smooth_window_ms}{An integer, length of smoothing window in ms. Default 20 ms (Liz's dissertation pg 121)}

\item{noise_reduction}{Boolean}

\item{...}{additional passed parameters from parent function}
}
\value{
data with smoothed gazepoint, distance, and pupilolmetry
}
\description{
De-noise and smooth rolling averages
}
