% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOffscreenGP.R
\name{detectOffscreenGaze}
\alias{detectOffscreenGaze}
\title{Detect Offscreen Gazepoints}
\usage{
detectOffscreenGaze(
  data,
  gazeX,
  gazeY,
  displayResolutionX_px,
  displayResolutionY_px,
  minDisplayResolutionX_px = 0,
  minDisplayResolutionY_px = 0
)
}
\arguments{
\item{data}{your dataframe}

\item{gazeX}{gaze point X coordinate}

\item{gazeY}{gaze point Y coordinate}

\item{displayResolutionX_px}{Integer, Display resolution X in pixels}

\item{displayResolutionY_px}{Integer Display resolution Y in pixels}

\item{minDisplayResolutionX_px}{Integer X minimum value within acceptable display resolution range Default 0 assumed for pixel space}

\item{minDisplayResolutionY_px}{Integer Y minimum value within acceptable display resolution range. Default 0 assumed for pixel space}
}
\value{
binary list to be saved to your dataframe as new column "offscreen.gp"
}
\description{
Detect Offscreen Gazepoints
}
