% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeSelect.r
\name{eyeSelect}
\alias{eyeSelect}
\title{Eye selection}
\usage{
eyeSelect(
  data,
  eyeSelection = "Maximize",
  gpLeft.X = "gazeLeftX.int",
  gpLeft.Y = "gazeLeftY.int",
  gpRight.X = "gazeRightX.int",
  gpRight.Y = "gazeRightY.int",
  pupilLeft = "pupilLeft.int",
  pupilRight = "pupilRight.int",
  distLeft.Z = "distanceLeftZ.int",
  distRight.Z = "distanceRightZ.int",
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{eyeSelection}{"Maximize", "Strict", "Left", or "Right"}

\item{gpLeft.X}{string column name for data containing left eye X-axis gazepoints}

\item{gpLeft.Y}{string column name for data containing left eye Y-axis gazepoints}

\item{gpRight.X}{string column name for data containing right eye X-axis gazepoints}

\item{gpRight.Y}{string column name for data containing right eye Y-axis gazepoints}

\item{pupilLeft}{string column name for data containing left eye pupil data}

\item{pupilRight}{string column name for data containing right eye pupil data}

\item{distLeft.Z}{string column name for data containing left eye distance from screen in mm (Z distance)}

\item{distRight.Z}{string column name for data containing right eye distance from screen in mm  (Z distance)}

\item{...}{additional passed parameters from parent function}
}
\value{
data
}
\description{
Eye selection
}
