% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGaps.r
\name{fillGaps}
\alias{fillGaps}
\title{Gap fill-in (interpolation)}
\usage{
fillGaps(data, recHz, intCols, max_gap_length = 50, ...)
}
\arguments{
\item{data}{dataframe}

\item{recHz}{recording Hz (calculated in calcHz)}

\item{intCols}{list of column names to be interpolated}

\item{max_gap_length}{integer for max ms of gaps to interpolate over. Default value 50ms (pg. 119 Liz's dissertation)}

\item{...}{additional passed parameters from parent function}
}
\value{
data with 8 new columns of interpolated left & right X/Y gaze points, left & right pupils, left and right Z position
}
\description{
Gap fill-in (interpolation)
}
