% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFixationIndicies.R
\name{findFixationIndices}
\alias{findFixationIndices}
\title{Find indices of fixations and saccades labelled by IVT filter}
\usage{
findFixationIndices(entry.class)
}
\arguments{
\item{entry.class}{a string specifying the classification column to use}
}
\value{
index summary - a list of fixation and saccade indexes, their start & end points
}
\description{
Find indices of fixations and saccades labelled by IVT filter
}
