% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEyeTrackingPlots.R
\name{generateEyeTrackingPlots}
\alias{generateEyeTrackingPlots}
\title{generateEyeTrackingPlots.R}
\usage{
generateEyeTrackingPlots(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
plots
}
\description{
generateEyeTrackingPlots.R
}
