% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_run_log_name.R
\name{get_file_run_log_name}
\alias{get_file_run_log_name}
\title{get_file_run_log_name}
\usage{
get_file_run_log_name(filename, batchName = NULL)
}
\arguments{
\item{filename}{filename}

\item{batchName}{text qualifier for the batch run}
}
\value{
list of ET derivatives data files to in BIDS-like directory
}
\description{
\code{get_file_run_log_name()} gets file run log name
}
