% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_difference.R
\name{get_time_difference}
\alias{get_time_difference}
\title{get_time_difference - get current time}
\usage{
get_time_difference(start, stop, units = "secs")
}
\arguments{
\item{start}{timestamp for start time}

\item{stop}{timestamp for stop time}

\item{units}{units difference should be returned, (secs, mins, hours, etc)}
}
\value{
timestamp in seconds, from sys.time()
}
\description{
get_time_difference - get current time
}
