% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileRunLogName.R
\name{getFileRunLogName}
\alias{getFileRunLogName}
\title{get_file_run_log_name}
\usage{
getFileRunLogName(filename, batchName = NULL)
}
\arguments{
\item{filename}{filename}

\item{batchName}{text qualifier for the batch run}
}
\value{
list of ET derivatives data files to in BIDS-like directory
}
\description{
\code{getFileRunLogName()} gets file run log name
}
