% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPipelineTiming.R
\name{getPipelineTiming}
\alias{getPipelineTiming}
\title{getPipelineTiming - functions to calculation processing time for pipeline}
\usage{
getPipelineTiming(start, stop)
}
\arguments{
\item{start}{timestamp for start time}

\item{stop}{timestamp for stop time}
}
\value{
time_diff a string stating the human readable elapsed time
}
\description{
getPipelineTiming - functions to calculation processing time for pipeline
}
