% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequenceGroupEndpoints.R
\name{getSequenceGroupEndpoints}
\alias{getSequenceGroupEndpoints}
\title{Get Sequence Group Endpoints}
\usage{
getSequenceGroupEndpoints(data, group_column)
}
\arguments{
\item{data}{dataframe}

\item{group_column}{name of column to group data over the event}
}
\value{
group_endpoints
}
\description{
Get Sequence Group Endpoints
}
