% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import Data}
\usage{
importData(filepath, ...)
}
\arguments{
\item{filepath}{path to data file}

\item{...}{additional passed parameters from parent function}
}
\value{
A tibble
}
\description{
Imports data from \code{csv}, \code{tsv}, and \code{xlsx} files
}
