% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivFlag.R
\name{ivFlag}
\alias{ivFlag}
\title{ivFlag
function ran from rmInvalidGP which marks invalid points as NA}
\usage{
ivFlag(data, eyeSelect, software, threshold = 2)
}
\arguments{
\item{data}{your dataset (processed by formatCols() function)}

\item{eyeSelect}{"left" or "right"}

\item{software}{either "TobiiStudio" or "TobiiPro"}

\item{threshold}{ONLY for Tobii Studio datasets. A numeric, indicating acceptable threshold for valid data.}
}
\value{
data with invalid points marked as NA in specified column
}
\description{
ivFlag
function ran from rmInvalidGP which marks invalid points as NA
}
