% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bids_derivative_files.R
\name{list_bids_derivative_files}
\alias{list_bids_derivative_files}
\title{list_bids_derivative_files}
\usage{
list_bids_derivative_files(
  directory,
  subject_pattern = "sub-[A-Z0-9]+",
  session_pattern = "ses-[0-9]+",
  derivatives_pattern = "derivatives",
  pipeline_pattern = "eyeQuality-v1",
  modality_pattern = NULL
)
}
\arguments{
\item{directory}{path for data file as .tsv}

\item{subject_pattern}{regex match pattern for subjects}

\item{session_pattern}{regex match pattern for sessions}

\item{derivatives_pattern}{regex match pattern for derivatives}

\item{pipeline_pattern}{regex match pattern for pipeline}

\item{modality_pattern}{regex match pattern for specific modality}
}
\value{
list of ET derivatives data files to in BIDS-like directory
}
\description{
\code{list_bids_derivative_files()} gets list of all ET like files in  derivatives.
}
