% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bids_files.R
\name{list_bids_files}
\alias{list_bids_files}
\title{list_bids_files - get list of all ET like files.}
\usage{
list_bids_files(
  directory,
  subject_pattern = "sub-[A-Z0-9]+",
  session_pattern = "ses-[0-9]+",
  modality_pattern = NULL
)
}
\arguments{
\item{directory}{path for data file as .tsv}

\item{subject_pattern}{regex match pattern for subjects}

\item{session_pattern}{regex match pattern for sessions}

\item{modality_pattern}{regex match pattern for specific modality}
}
\value{
list of raw ET data files to in BIDS-like directory
}
\description{
list_bids_files - get list of all ET like files.
}
