% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBidsDerivativeFiles.R
\name{listBidsDerivativeFiles}
\alias{listBidsDerivativeFiles}
\title{listBidsDerivativeFiles}
\usage{
listBidsDerivativeFiles(
  directory,
  subjectPattern_regex = "sub-[A-Z0-9]+",
  sessionPattern_regex = "ses-[0-9]+",
  modalityPattern_regex = NULL,
  recursiveSearch = FALSE,
  derivativesDirectory_name = "derivatives",
  pipelineDirectory_name = "eyeQuality-v1",
  derivativePattern_regex = NULL,
  ...
)
}
\arguments{
\item{directory}{path for data file as .tsv}

\item{subjectPattern_regex}{regex match pattern for subjects}

\item{sessionPattern_regex}{regex match pattern for sessions}

\item{modalityPattern_regex}{regex match pattern for specific modality}

\item{recursiveSearch}{boolean to search file directory recursively}

\item{derivativesDirectory_name}{name for derivatives directory, default 'derivatives'}

\item{pipelineDirectory_name}{name for pipeline processing directory, default 'eyeQuality-v1'}

\item{derivativePattern_regex}{regex match pattern for specific derivative files}

\item{...}{additional parameters that may get passed from wrapper functions}
}
\value{
list of ET derivatives data files to in BIDS-like directory
}
\description{
\code{listBidsDerivativeFiles()} gets list of all ET like files in  derivatives.
}
