% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapData.R
\name{mapData}
\alias{mapData}
\title{map column names from data input}
\usage{
mapData(data, software)
}
\arguments{
\item{data}{dataframe}

\item{software}{"TobiiStudio" or "TobiiPro"}
}
\value{
your reformatted data with standardized columns
}
\description{
To account for varying column names between data export types we rename columns to the following standard set:
(event, eventValue, recordingDuration_ms, resolutionHeight, resolutionWidth, eyeTrackerTimestamp,recordingTimestamp_ms, gazeLeftX, gazeLeftY, gazeRightX, gazeRightY, distanceLeftZ, distanceRightZ, pupilLeft, pupilRight, validityLeft, validityRight)
}
