% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAdjacentFixation.R
\name{mergeAdjacentFixations}
\alias{mergeAdjacentFixations}
\title{Merge Adjacent Fixations}
\usage{
mergeAdjacentFixations(
  data,
  entryClass,
  rleFixIndex,
  gazeX,
  gazeY,
  fixationEnd,
  fixationStart,
  fixationLength,
  mergeDistance_va = 0.5,
  mergeTimeGap_ms = 75,
  sampling_interval = 3.3
)
}
\arguments{
\item{data}{dataframe}

\item{entryClass}{list of IVT-assigned class assignments per gazepoint}

\item{rleFixIndex}{list of integers, generated from output of function findFixationIndices}

\item{gazeX}{list of x gaze coordinates as visual angles}

\item{gazeY}{list of y gaze coordinates as visual angles}

\item{fixationEnd}{list of fixation endpoints as indexes, generated from output of function findFixationIndices}

\item{fixationStart}{list of fixation start points as indexes, generated from output of function findFixationIndices}

\item{fixationLength}{list of fixation durations, generated from output of function findFixationIndices}

\item{mergeDistance_va}{maximum distance (in VA) between two fixations to be merged. Default: 0/5 VA}

\item{mergeTimeGap_ms}{maximum time lapsed (in ms) between two fixations to be merged. Default: 75 ms}

\item{sampling_interval}{recording's actual sample-to-sample interval, in ms (i.e. 1000 / recordingFrequency_hz). Used to convert a gap duration in samples to ms when comparing against mergeTimeGap_ms. Default: 3.3, matching a ~303 Hz recording}
}
\value{
return list of updated fixation metrics, including classification (class.adj), duration, and distance
}
\description{
Merge Adjacent Fixations
}
