% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naFlag.R
\name{naFlag}
\alias{naFlag}
\title{naFlag
changes datapoints marked as -9999 in TobiiStudio to NA}
\usage{
naFlag(data, software)
}
\arguments{
\item{data}{a dataframe}

\item{software}{"TobiiPro" or "TobiiStudio}
}
\value{
data with recoded NA values
}
\description{
naFlag
changes datapoints marked as -9999 in TobiiStudio to NA
}
