% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offscreenCategory.R
\name{offscreenCategory}
\alias{offscreenCategory}
\title{Categorize offScreen gazepoints based on eye selection}
\usage{
offscreenCategory(
  eyeSelection,
  gazeLeft.offscreen,
  gazeRight.offscreen,
  offscreen.eyeSelect
)
}
\arguments{
\item{eyeSelection}{"Maximize", "Strict", "Left", "Right"}

\item{gazeLeft.offscreen}{list of ordered gaze point offscreen assignments for left eye}

\item{gazeRight.offscreen}{list of ordered gaze point offscreen assignments for right eye}

\item{offscreen.eyeSelect}{list of ordered gaze point offscreen assignments for eyeSelect column}
}
\value{
categoryList
}
\description{
Categorize offScreen gazepoints based on eye selection
}
