% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_timing.R
\name{pipeline_timing}
\alias{pipeline_timing}
\title{pipeline_timing - functions to calculation processing time for pipeline}
\usage{
pipeline_timing(start, stop)
}
\arguments{
\item{start}{timestamp for start time}

\item{stop}{timestamp for stop time}
}
\value{
time_diff a string stating the human readable elapsed time
}
\description{
pipeline_timing - functions to calculation processing time for pipeline
}
