% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGazeAndBlinks.R
\name{plotGazeAndBlinks}
\alias{plotGazeAndBlinks}
\title{plotGazeAndBlinks}
\usage{
plotGazeAndBlinks(data, column, show_fixations = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{column}{name of data column you want plotted as points/lines}

\item{show_fixations}{boolean TRUE/FALSE to indicate if boxes of color to show fixations should be plotted}
}
\value{
plots
}
\description{
plotGazeAndBlinks
}
