% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeIntermediateCols.R
\name{removeIntermediateCols}
\alias{removeIntermediateCols}
\title{Remove columns calculated during intermediate steps of run_preprocess}
\usage{
removeIntermediateCols(data, ...)
}
\arguments{
\item{data}{dataframe containing preprocessed data}

\item{...}{additional passed parameters from parent function}
}
\value{
data
}
\description{
Remove columns calculated during intermediate steps of run_preprocess
}
