% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameColumns.R
\name{renameColumns}
\alias{renameColumns}
\title{rename column names to standard format from data input based on recording software}
\usage{
renameColumns(data, software)
}
\arguments{
\item{data}{dataframe}

\item{software}{"TobiiStudio" or "TobiiPro"}
}
\value{
your reformatted data with standardized columns
}
\description{
To account for varying column names between data export types we rename columns to the following standard set:
(event, eventValue, recordingDuration_ms, resolutionHeight, resolutionWidth, eyeTrackerTimestamp,recordingTimestamp_ms, gazeLeftX, gazeLeftY, gazeRightX, gazeRightY, distanceLeftZ, distanceRightZ, pupilLeft, pupilRight, validityLeft, validityRight)
}
