% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmInvalid.r
\name{rmInvalidGP}
\alias{rmInvalidGP}
\title{Remove invalid gazepoints}
\usage{
rmInvalidGP(data, software, threshold = 2, ...)
}
\arguments{
\item{data}{your eyetracking data (a dataframe)}

\item{software}{Eye tracking software source, "TobiiStudio" or "TobiiPro"}

\item{threshold}{a numeric (0, 1, 2, 3, or 4) - Only applicable if software = "TobiiStudio"}

\item{...}{additional passed parameters from parent function}
}
\value{
your processed data, with Invalid GP marked as NA
}
\description{
Remove invalid gazepoints
}
\examples{
#mydataset <- tobii_studio_example_data
#rmInvalidGP(mydataset, "TobiiStudio", threshold = 2)
#mydataset2 <- tobii_pro_example_data
#rmInvalidGP(mydataset2, "TobiiPro")
}
