% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmOffscreenGP.R
\name{rmOffscreenGP}
\alias{rmOffscreenGP}
\title{Remove offscreen gazepoints and related interpolations - rename to excludeOffscreenGazepoint?}
\usage{
rmOffscreenGP(
  data,
  gazeX,
  gazeY,
  distZ,
  overwrite = NA,
  display.resx,
  display.resy,
  display.dimx,
  display.dimy,
  valid_offscreen = 5,
  ...
)
}
\arguments{
\item{data}{your dataframe}

\item{gazeX}{string column containing name of variable with X gaze data}

\item{gazeY}{string column containing name of variable with Y gaze data}

\item{distZ}{string column containing name of variable with Z distance data in mm}

\item{overwrite}{optional list of column names in addition to gaze columns which should be replaced by NA when offscreen. Default = NA}

\item{display.resx}{X Display resolution in pixels}

\item{display.resy}{Y Display resolution in pixels}

\item{display.dimx}{X Display dimension in mm}

\item{display.dimy}{Y Display dimension in mm}

\item{valid_offscreen}{integer Degrees of Visual angle outside of screen which can be kept as valid. Default = 5 VA}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
data
}
\description{
Remove offscreen gazepoints and related interpolations - rename to excludeOffscreenGazepoint?
}
