% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_preprocess.R
\name{run_preprocess}
\alias{run_preprocess}
\title{Wrapper - run all preprocessing functions}
\usage{
run_preprocess(
  filepath,
  display.dimx,
  display.dimy,
  data = NULL,
  eyeSelection = "Maximize",
  noise_reduction = TRUE,
  savedata = FALSE,
  includeIntermediates = FALSE,
  firstEvent = NULL,
  lastEvent = NULL,
  timeStart = NULL,
  timeEnd = NULL,
  batchName = NULL,
  ...
)
}
\arguments{
\item{filepath}{path for data file as .tsv}

\item{display.dimx}{integer of display width in millimeters. For example our 1920x1080 screen has a width of 594 mm}

\item{display.dimy}{integer of display height in millimeters. For example our 1920x1080 screen has a width of 344 mm}

\item{data}{instead of passing a filepath, you can pass a data frame with your loaded data. Default NULL}

\item{eyeSelection}{string with possible values "Maximize", "Strict", "Left", or "Right". Default "Maximize"}

\item{noise_reduction}{Boolean indicating if we should run denoise script}

\item{savedata}{Boolean indicating if we should suppress output, and save data to files}

\item{includeIntermediates}{Boolean indicating if column outputs from all intermediate steps of the preprocessing should be included. Default = FALSE}

\item{firstEvent}{optional String of the first event marker to include. Default NULL}

\item{lastEvent}{optional String of the last event marker to include. Default NULL}

\item{timeStart}{optional Recording Timestamp of the first data point to include. Default NULL}

\item{timeEnd}{optional Recording Timestamp of the last data point to include. Default NULL}

\item{batchName}{optional string to append output files with a specific batch run label. Default NULL}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
preprocessed data
}
\description{
Wrapper - run all preprocessing functions
}
