% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTimeRange.R
\name{setTimestamps}
\alias{setTimestamps}
\title{Set time range of data to be processed}
\usage{
setTimestamps(data, timeStart, timeEnd, setTimes = TRUE, ...)
}
\arguments{
\item{data}{your dataframe of eyetracking data}

\item{timeStart}{an integer of recording timestamp where data of interest begins}

\item{timeEnd}{an integer of recording timestamp where data of interest ends}

\item{setTimes}{boolean TRUE/FALSE if we should filter data based on the timeStart and timeEnd provided}

\item{...}{arguments to be passed}
}
\value{
dataframe containing only data within the specified timeframe
}
\description{
Set time range of data to be processed
}
