% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFiles.R
\name{sinkToOutputFile}
\alias{sinkToOutputFile}
\title{Set sink cmd outputs to save to run log output file}
\usage{
sinkToOutputFile(runlog)
}
\arguments{
\item{runlog}{string of filepath to run log}
}
\description{
Set sink cmd outputs to save to run log output file
}
