% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothGaze.R
\name{smoothGaze}
\alias{smoothGaze}
\title{De-noise and smooth rolling averages}
\usage{
smoothGaze(
  data,
  recordingFrequency_hz,
  columnsToSmooth,
  smoothingWindow_ms = 20,
  smoothGaze_boolean = TRUE,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{recordingFrequency_hz}{recording Hz}

\item{columnsToSmooth}{list of columns to smooth. Recommended to smooth gaze, pupil, and Z Dist columns}

\item{smoothingWindow_ms}{An integer, length of smoothing window in ms. Default 20 ms (Liz's dissertation pg 121)}

\item{smoothGaze_boolean}{Boolean}

\item{...}{additional passed parameters from parent function}
}
\value{
data with smoothed gazepoint, distance, and pupillometry
}
\description{
De-noise and smooth rolling averages
}
