% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothVel.r
\name{smoothVel}
\alias{smoothVel}
\title{Smooth Velocity}
\usage{
smoothVel(data, recHz, velCols, vel_window_ms = 10, ...)
}
\arguments{
\item{data}{dataframe}

\item{recHz}{recording Hz}

\item{velCols}{list of column names which contain velocity data to be smoothed. These columns should already be converted from pixels to VA}

\item{vel_window_ms}{An integer. Velocity smoothing window size in milliseconds. Default: 10ms (pg 121 Liz's dissertation)}

\item{...}{additional passed parameters from parent function}
}
\value{
data with smoothed velocity column
}
\description{
Smooth Velocity
}
