% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothVelocity.R
\name{smoothVelocity}
\alias{smoothVelocity}
\title{Smooth Velocity}
\usage{
smoothVelocity(
  data,
  recordingFrequency_hz,
  velocity,
  smoothVelocityWindow_ms = 10,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{recordingFrequency_hz}{recording Hz}

\item{velocity}{list of column names which contain velocity data to be smoothed. These columns should already be converted from pixels to VA}

\item{smoothVelocityWindow_ms}{An integer. Velocity smoothing window size in milliseconds. Default: 10ms (pg 121 Liz's dissertation)}

\item{...}{additional passed parameters from parent function}
}
\value{
data with smoothed velocity column
}
\description{
Smooth Velocity
}
