% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCols.R
\name{standardizeColumnNames}
\alias{standardizeColumnNames}
\title{Run renameColumns or direct to custom mapping}
\usage{
standardizeColumnNames(data, software)
}
\arguments{
\item{data}{dataframe}

\item{software}{"TobiiStudio" or "TobiiPro"}
}
\value{
your reformatted data with standardized column names (Event, EventValue, RecordingDuration, ResolutionHeight, ResolutionWidth, EyeTrackerTimestamp, RecordingTimestamp, GazePointLeftX, GazePointLeftY, GazePointRightX, GazePointRightY, EyePosLeftX, EyePosLeftY, EyePosLeftZ, EyePosRightX, EyePosRightY, EyePosRightZ, PupilLeft, PupilRight, ValidityLeft, ValidityRight)
}
\description{
Run renameColumns or direct to custom mapping
}
