% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va2pixels.R
\name{va2pixels}
\alias{va2pixels}
\title{Calculate pixel position from visual angle}
\usage{
va2pixels(va, distZ, display.res, display.dim)
}
\arguments{
\item{va}{integer Visual angle to be converted to pixel space}

\item{distZ}{integer distance Z from the screen in mm}

\item{display.res}{display resolution in px}

\item{display.dim}{display dimension in mm}
}
\value{
px the pixel value for the input VA
}
\description{
Calculate pixel position from visual angle
}
