% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visAngle.r
\name{visAngle}
\alias{visAngle}
\title{Calculate Visual Angle}
\usage{
visAngle(pospx, posz, disp.res, disp.dim)
}
\arguments{
\item{pospx}{Integer, gaze position as a pixel coordinate (X or Y)}

\item{posz}{Integer, distance from the screen in mm}

\item{disp.res}{Integer, display resolution on the relevant dimension (X ir Y)}

\item{disp.dim}{Integer, display dimension on the relevant dimension (X or Y)}
}
\value{
returns Angle in degrees
}
\description{
Calculate Visual Angle
}
