% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOutputMetrics.R
\name{calculateOutputMetrics}
\alias{calculateOutputMetrics}
\title{Calculate Output Metrics}
\usage{
calculateOutputMetrics(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
summary_df dataframe describing each
}
\description{
Calculate Output Metrics
}
