% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVisualAngle.R
\name{calculateVisualAngle}
\alias{calculateVisualAngle}
\title{Calculate Visual Angle from pixel position}
\usage{
calculateVisualAngle(
  gaze_px,
  distanceZ,
  displayResolution_px,
  displayDimension_mm
)
}
\arguments{
\item{gaze_px}{Integer, gaze position as a pixel coordinate (X or Y)}

\item{distanceZ}{Integer, distance from the screen in mm}

\item{displayResolution_px}{Integer, display resolution on the relevant dimension (X or Y)}

\item{displayDimension_mm}{Integer, display dimension on the relevant dimension (X or Y)}
}
\value{
returns Angle in degrees
}
\description{
Calculate Visual Angle from pixel position
}
