% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyBlinks.R
\name{classifyBlinks}
\alias{classifyBlinks}
\title{Function to detect blinks adapted from a noise-based algorithm developed by Hershman, Henik, & Cohen (2018) https://link.springer.com/article/10.3758/s13428-017-1008-1; https://osf.io/jyz43/}
\usage{
classifyBlinks(
  data,
  pupilLeft = "pupilLeft.int",
  pupilRight = "pupilRight.int",
  recordingFrequency_hz,
  maxArtifactLength_ms = 15,
  minBlinkLength_ms = 100,
  maxBlinkLength_ms = 400,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{pupilLeft}{name of columns containing left pupil data to be analyzed}

\item{pupilRight}{name of columns containing right pupil data to be analyzed}

\item{recordingFrequency_hz}{integer from calculateFrequency_hz.R}

\item{maxArtifactLength_ms}{integer maximum length of valid data segment to be considered an artifact}

\item{minBlinkLength_ms}{integer min value for missing that could be considered a blink}

\item{maxBlinkLength_ms}{integer max value for missing that could be considered a blink}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
dataframe with new columns indicating presence/absence of blinks for each pupil data type
}
\description{
Function to detect blinks adapted from a noise-based algorithm developed by Hershman, Henik, & Cohen (2018) https://link.springer.com/article/10.3758/s13428-017-1008-1; https://osf.io/jyz43/
}
