% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVisualAngToPixels.R
\name{convertVisualAngToPixels}
\alias{convertVisualAngToPixels}
\title{Calculate pixel position from visual angle}
\usage{
convertVisualAngToPixels(
  gazeVA,
  distanceZ,
  displayResolution_px,
  displayDimension_mm
)
}
\arguments{
\item{gazeVA}{integer Visual angle to be converted to pixel space}

\item{distanceZ}{integer distance Z from the screen in mm}

\item{displayResolution_px}{display resolution in px}

\item{displayDimension_mm}{display dimension in mm}
}
\value{
px the pixel value for the input VA
}
\description{
Calculate pixel position from visual angle
}
