% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFiles.R
\name{create_new_filename}
\alias{create_new_filename}
\title{create_new_filename: create new output filename from save directory and data file}
\usage{
create_new_filename(inputfile, appendname, newFileExtension = NULL)
}
\arguments{
\item{inputfile}{name of input file}

\item{appendname}{text to append before the file extension}

\item{newFileExtension}{new file extension to use, if different from filename}
}
\description{
create_new_filename: create new output filename from save directory and data file
}
