% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectImportSourceType.R
\name{detectImportSourceType}
\alias{detectImportSourceType}
\title{Detect Import Source Type}
\usage{
detectImportSourceType(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
A string
}
\description{
Detects if imported data comes from TobiiPro, TobbiStudio, or another Eye Tracking export type
}
