% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetrackingELabR-package.R
\docType{package}
\name{eyeQuality-package}
\alias{eyeQuality}
\alias{eyeQuality-package}
\title{eyeQuality: What the Package Does (One Line, Title Case)}
\description{
Preprocessing R package for transparency, reproducibility, and measurement of data quality in eye tracking datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elab-umn/eyeQuality}
  \item Report bugs at \url{https://github.com/elab-umn/eyeQuality/issues}
}

}
\author{
\strong{Maintainer}: Aaron Glick \email{glick094@umn.edu} (\href{https://orcid.org/0000-0001-6926-7166}{ORCID})

Authors:
\itemize{
  \item Lana Hantzsch \email{hantz006@umn.edu}
  \item Eli Johnson
}

Other contributors:
\itemize{
  \item Elizabeth Sharer [contributor]
  \item Caitlin Sisk [contributor]
  \item Isa Stallworthy [contributor]
}

}
\keyword{internal}
