% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeQuality.R
\name{eyeQuality}
\alias{eyeQuality}
\title{eyeQuality wrapper - run all preprocessing functions}
\usage{
eyeQuality(
  filepath,
  displayDimensionX_mm,
  displayDimensionY_mm,
  data = NULL,
  eyeSelection_method = "Maximize",
  smoothGaze_boolean = TRUE,
  maxValidityThreshold = 2,
  saveData = FALSE,
  includeIntermediates = FALSE,
  studioEvents = NULL,
  proEvents = NULL,
  timeStart = NULL,
  timeEnd = NULL,
  batchName = NULL,
  ...
)
}
\arguments{
\item{filepath}{path for data file as .tsv}

\item{displayDimensionX_mm}{integer of display width in millimeters. For example our 1920x1080 screen has a width of 594 mm}

\item{displayDimensionY_mm}{integer of display height in millimeters. For example our 1920x1080 screen has a width of 344 mm}

\item{data}{instead of passing a filepath, you can pass a data frame with your loaded data. Default NULL}

\item{eyeSelection_method}{string with possible values "Maximize", "Strict", "Left", or "Right". Default "Maximize"}

\item{smoothGaze_boolean}{Boolean indicating if we should run smoothGaze script}

\item{maxValidityThreshold}{a numeric (0, 1, 2, 3, or 4) describing what validity rating is considered acceptable - Only applicable if software = "TobiiStudio"}

\item{saveData}{Boolean indicating if we should suppress output, and save data to files}

\item{includeIntermediates}{Boolean indicating if column outputs from all intermediate steps of the preprocessing should be included. Default = FALSE}

\item{studioEvents}{optional list of two values specifying the start and ending event labels for Tobii Studio files, c(startEventName, endEventName) . Default NULL}

\item{proEvents}{optional list of two values specifying the start and ending event labels for Tobii Pro files, c(startEventName, endEventName) . Default NULL}

\item{timeStart}{optional Recording Timestamp of the first data point to include. Default NULL}

\item{timeEnd}{optional Recording Timestamp of the last data point to include. Default NULL}

\item{batchName}{optional string to append output files with a specific batch run label. Default NULL}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
preprocessed data
}
\description{
eyeQuality wrapper - run all preprocessing functions
}
