% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeSelect.r
\name{eyeSelect}
\alias{eyeSelect}
\title{Eye selection}
\usage{
eyeSelect(
  data,
  eyeSelection_method = "Maximize",
  gazeLeftX = "gazeLeftX.int",
  gazeLeftY = "gazeLeftY.int",
  gazeRightX = "gazeRightX.int",
  gazeRightY = "gazeRightY.int",
  pupilLeft = "pupilLeft.int",
  pupilRight = "pupilRight.int",
  distanceLeftZ = "distanceLeftZ.int",
  distanceRightZ = "distanceRightZ.int",
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{eyeSelection_method}{"Maximize", "Strict", "Left", or "Right"}

\item{gazeLeftX}{string column name for data containing left eye X-axis gazepoints}

\item{gazeLeftY}{string column name for data containing left eye Y-axis gazepoints}

\item{gazeRightX}{string column name for data containing right eye X-axis gazepoints}

\item{gazeRightY}{string column name for data containing right eye Y-axis gazepoints}

\item{pupilLeft}{string column name for data containing left eye pupil data}

\item{pupilRight}{string column name for data containing right eye pupil data}

\item{distanceLeftZ}{string column name for data containing left eye distance from screen in mm (Z distance)}

\item{distanceRightZ}{string column name for data containing right eye distance from screen in mm  (Z distance)}

\item{...}{additional passed parameters from parent function}
}
\value{
data
}
\description{
Eye selection
}
