% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{getCurrentTime}
\alias{getCurrentTime}
\title{getCurrentTime - get current time}
\usage{
getCurrentTime()
}
\value{
timestamp in seconds, from sys.time()
}
\description{
getCurrentTime - get current time
}
