% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateGaze.R
\name{interpolateGaze}
\alias{interpolateGaze}
\title{Interpolate gaps in gaze data}
\usage{
interpolateGaze(
  data,
  recordingFrequency_hz,
  columnsToInterpolate,
  maxGapLength_ms = 50,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{recordingFrequency_hz}{recording Hz (calculated in calculateFrequency_hz)}

\item{columnsToInterpolate}{list of column names to be interpolated}

\item{maxGapLength_ms}{integer for max ms of gaps to interpolate over. Default value 50ms (pg. 119 Liz's dissertation)}

\item{...}{additional passed parameters from parent function}
}
\value{
data with 8 new columns of interpolated left & right X/Y gaze points, left & right pupils, left and right Z position
}
\description{
Interpolate gaps in gaze data
}
