% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBidsDerivativeFiles.R
\name{listBidsDerivativeFiles}
\alias{listBidsDerivativeFiles}
\title{listBidsDerivativeFiles}
\usage{
listBidsDerivativeFiles(
  directory,
  subjectPattern_regex = "sub-[A-Z0-9]+",
  session_pattern = "ses-[0-9]+",
  derivativesPattern_regex = "derivatives",
  pipeline_pattern = "eyeQuality-v1",
  modalityPattern_regex = NULL
)
}
\arguments{
\item{directory}{path for data file as .tsv}

\item{subjectPattern_regex}{regex match pattern for subjects}

\item{derivativesPattern_regex}{regex match pattern for derivatives}

\item{pipeline_pattern}{regex match pattern for pipeline}

\item{modalityPattern_regex}{regex match pattern for specific modality}

\item{sessionPattern_regex}{regex match pattern for sessions}
}
\value{
list of ET derivatives data files to in BIDS-like directory
}
\description{
\code{listBidsDerivativeFiles()} gets list of all ET like files in  derivatives.
}
