% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePreprocessingBatchSummary.R
\name{parsePreprocessingBatchSummary}
\alias{parsePreprocessingBatchSummary}
\title{parsePreprocessingBatchSummary}
\usage{
parsePreprocessingBatchSummary(batch_summary_file, info_to_extract = "summary")
}
\arguments{
\item{batch_summary_file}{filepath to preprocessing_batch_summary.tsv file}

\item{info_to_extract}{string with value of "summary" | "failedfiles" | "successfulfiles"}
}
\value{
data dataframe exported data from preprocessing_batch_summary
}
\description{
parse preprocessing_batch_summary files
}
