% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmOffscreenGP.R
\name{removeOffscreenGaze}
\alias{removeOffscreenGaze}
\title{Remove offscreen gazepoints and related interpolations - rename to excludeOffscreenGazepoint?}
\usage{
removeOffscreenGaze(
  data,
  gazeX,
  gazeY,
  distanceZ,
  overwrite = NA,
  displayResolutionX_px,
  displayResolutionY_px,
  displayDimensionX_mm,
  displayDimensionY_mm,
  offscreenValidityRange_va = 5,
  ...
)
}
\arguments{
\item{data}{your dataframe}

\item{gazeX}{string column containing name of variable with X gaze data}

\item{gazeY}{string column containing name of variable with Y gaze data}

\item{distanceZ}{string column containing name of variable with Z distance data in mm}

\item{overwrite}{optional list of column names in addition to gaze columns which should be replaced by NA when offscreen. Default = NA}

\item{displayResolutionX_px}{X Display resolution in pixels}

\item{displayResolutionY_px}{Y Display resolution in pixels}

\item{displayDimensionX_mm}{X Display dimension in mm}

\item{displayDimensionY_mm}{Y Display dimension in mm}

\item{offscreenValidityRange_va}{integer Degrees of Visual angle outside of screen which can be kept as valid. Default = 5 VA}

\item{...}{additional arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
data
}
\description{
Remove offscreen gazepoints and related interpolations - rename to excludeOffscreenGazepoint?
}
