% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeShortFixations.R
\name{removeShortFixations}
\alias{removeShortFixations}
\title{Removes proposed fixations that are too short}
\usage{
removeShortFixations(
  data,
  entryClass,
  rleFixIndex,
  fixationEnd,
  fixationStart,
  fixationLength,
  shortFixationThreshold_ms
)
}
\arguments{
\item{data}{dataframe}

\item{entryClass}{label for entry classification}

\item{rleFixIndex}{list of integers, generated from output of function findFixationIndices}

\item{fixationEnd}{list of fixation endpoints as indexes, generated from output of function findFixationIndices}

\item{fixationStart}{list of fixation start points as indexes, generated from output of function findFixationIndices}

\item{fixationLength}{integer list of fixation durations, generated from output of function findFixationIndices}

\item{shortFixationThreshold_ms}{threshold for minimum duration of a given fixation to be considered acceptable. Recommended default: 60 ms}
}
\value{
return list of updated fixation metrics, including classification (class.adj), duration, and distance
}
\description{
Removes proposed fixations that are too short
}
