% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFiles.R
\name{saveFiles}
\alias{saveFiles}
\title{Save Files}
\usage{
saveFiles(inputFile, data, events, timing, summaryData, batchName = NULL)
}
\arguments{
\item{inputFile}{string filepath of the input file, to generate save file names}

\item{data}{dataframe}

\item{events}{dataframe of event markers, only tobii pro files}

\item{timing}{list of internal run timing from eyeQuality function}

\item{summaryData}{data from the calcualteOutputMetrics function}

\item{batchName}{batch name to insert into save files, useful for running batches with different parameters and/or for specific trials}
}
\description{
Save Files
}
