% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.r
\name{smoothGaze}
\alias{smoothGaze}
\title{De-noise and smooth rolling averages}
\usage{
smoothGaze(
  data,
  recordingFrequency_hz,
  columnsToSmooth,
  smoothingWindow_ms = 20,
  smoothGaze_boolean = TRUE,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{...}{additional passed parameters from parent function}

\item{recHz}{recording Hz}

\item{input_cols}{list of columns to smooth. Recommended to smooth gaze, pupil, and Z Dist columns}

\item{smooth_window_ms}{An integer, length of smoothing window in ms. Default 20 ms (Liz's dissertation pg 121)}

\item{noise_reduction}{Boolean}
}
\value{
data with smoothed gazepoint, distance, and pupilolmetry
}
\description{
De-noise and smooth rolling averages
}
